training/requirements.txt
# Training Dependencies

# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
peft>=0.7.0

# Evaluation
clip-by-openai>=1.0
torchmetrics>=1.2.0
torch-fidelity>=0.3.0

# Image Processing
Pillow>=10.0.0

# Utilities
tqdm>=4.66.0
numpy>=1.24.0
matplotlib>=3.7.0
datasets>=2.14.0

# Model Hub
huggingface-hub>=0.19.0
safetensors>=0.4.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Optional: Performance
xformers>=0.0.22
bitsandbytes>=0.41.0  # For 8-bit optimization